requirements.txt
opencv-python
bettercam
Pillow
pynput